+++ src/main/java/org/cinema/model/User.java
@@ -13,10 +13,9 @@ import java.time.ZoneId;
 @AllArgsConstructor
 @NoArgsConstructor
 public class User {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false, unique = true)
     private String username;
